feat: seo generation for grant and projects
